Sukurtas atskiras .h failas šabloninėm fukcijom laikyti
